杀进程程序会杀自己，然后不工作

+++ mylinux/rm_ps.py
@@ -2,10 +2,14 @@
 import sys
 import os
 import re
-tmp=os.popen(r'ps').readlines()
+tmp=os.popen(r'ps -t').readlines()
 for i in tmp:
-	if re.search("perl",i) or re.search("python",i) or re.search("spectre",i) or re.search(r"awd.exe",i):
-		aaa=re.split("\s*[p]",i)[0]
-		bbb=aaa.split(" ")[-1]
-		os.system('kill -9 ' + str(bbb))
-
+	if re.search(r"rm_ps",i):
+		pass
+	else:
+		if re.search("perl",i) or re.search("python",i) or re.search("spectre",i) or re.search(r"awd.exe",i):
+			#print i
+			aaa=re.split("\s*[p]",i)[0]
+			bbb=aaa.split(" ")[-1]
+			#print "pid=",bbb
+			os.system('kill -9 ' + str(bbb))
